fix(p5):auth helpers util & stp decimal truncation

+++ backend/src/main/java/com/senior/spm/service/dto/JiraIssueDto.java
@@ -3,7 +3,7 @@ package com.senior.spm.service.dto;
 public record JiraIssueDto(
         String issueKey,
         String assigneeEmail,
-        Integer storyPoints,
+        Double storyPoints,
         String description
 ) {
 }
